Completed most endpoints, updated user models to work with Shibboleth, added endtime to blips, time is now stored as a DATE field

+++ app/models/Group.scala
@@ -3,8 +3,8 @@ package models
 import scala.collection.mutable.MutableList
 
 case class Group(
-    gid: Int,
-    name: String, 
+    gid: Long,
+    name: String,
     concrete: Boolean,
     interests: List[Long] = List[Long]()
 )
